requirements.txt
asgiref
Django
django-extensions
djangorestframework
pillow
sqlparse
tzdata
sqlalchemy
faker
dotenv
numpy 
pandas
seaborn
openai